added argv parsing

+++ utils/call_backs.c
@@ -6,11 +6,11 @@
 /*   By: jkovacev <jkovacev@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/20 20:27:32 by jkovacev          #+#    #+#             */
-/*   Updated: 2025/06/27 17:54:07 by jkovacev         ###   ########.fr       */
+/*   Updated: 2025/06/27 22:36:43 by jkovacev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
+#include "../minishell.h"
 
 void	delete_env_var(void *var)
 {
@@ -31,7 +31,7 @@ bool	get_env_value(void *content, void *var_key)
 	var = (t_env_var *)content;
 	k = (char *)var_key;
 	len = ft_strlen(var_key);
-	if (len != ft_strlen(var->key))
+	if (len != (int)ft_strlen(var->key))
 		return (false);
 	return (ft_strncmp(var->key, k, len) == 0);
 }
